learning how to authenticate requests

+++ TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
 from models import Todos
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -31,16 +33,24 @@ that will allow us to be able to do some kind of code behing the scenes and then
 in this case the dependency injection is Depends(get_db) which makes our DB now have a session local
 """
 @router.get("/",  status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 """
 fetching a single todo
 """
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+    .filter(Todos.owner_id == user.get('id')).first()
+    
+    
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found buddy.')
@@ -49,8 +59,13 @@ async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
 create and save todos
 """
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency,
+                      db: db_dependency,
+                      todo_request: TodoRequest):
+    
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
     
     db.add(todo_model)
     db.commit()
@@ -59,11 +74,16 @@ async def create_todo(db: db_dependency, todo_request: TodoRequest):
 update todo (put request method)
 """
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+async def update_todo(user: user_dependency,
+                      db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0) ):
     
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+    .filter(Todos.owner_id == user.get('id')).first()
     
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found buddy.')
@@ -80,11 +100,17 @@ async def update_todo(db: db_dependency,
 delete todo (delete request method)
 """
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency,
+                      db: db_dependency, todo_id: int = Path(gt=0)):
+    
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+    .filter(Todos.owner_id == user.get('id')).first()
     
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found buddy.')
     
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
